Refact: made the Data object one big class.
